[DuyLP]

+++ src/main/java/com/dinh/logistics/respository/CollectPointRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CollectPointRepository extends JpaRepository<CollectPoint, Long> {
+public interface CollectPointRepository extends JpaRepository<CollectPoint, Integer> {
 }
