+++ src/main/java/com/dinh/logistics/respository/MediaRepository.java
@@ -3,5 +3,5 @@ package com.dinh.logistics.respository;
 import com.dinh.logistics.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MediaRepository extends JpaRepository<Employee, Long> {
+public interface MediaRepository extends JpaRepository<Employee, Integer> {
 }
